Reformatted Readers to EmpaticaReader
